requirements/prod.txt
-r base.txt
django-postgrespool==0.3.0
gunicorn==19.3.0
